Better python check, expose math and more
