+++ examples/example_004.py
@@ -1,10 +1,14 @@
 from pyperlib import *
 
-
 def start():
     print("Start from example 4!")
 
 
 def update():
     fill(colors.dark_green)
+    print(PI)
+    print(E**2)
+    print(TAU)
+    print(sin(PI))
+    print(tan(TAU))
     print("Update from example 4!")
