+++ pyperlib/__main__.py
@@ -1,7 +1,15 @@
 import sys
 import argparse
+import time
+from os import path, getcwd
+from importlib import util as import_util
 
-from pyperlib.drawing.window import run_pyper_window
+import moderngl
+import pyglet
+
+from pyperlib.settings import settings
+from pyperlib.drawing.window import PyperWindow
+
 
 def cli():
     try:
@@ -26,6 +34,38 @@ def cli():
     except argparse.ArgumentError as error:
         print(str(error))
         sys.exit(2)
+
+
+def run_pyper_window(arguments):
+    file_path = path.join(getcwd(), arguments.filename)
+    spec = import_util.spec_from_file_location("", file_path)
+    user_sketch = import_util.module_from_spec(spec)
+    spec.loader.exec_module(user_sketch)
+
+    window = PyperWindow(
+        resizable=True,
+        fullscreen=arguments.fullscreen
+    )
+
+    context = moderngl.create_context(require=settings.opengl_version)
+
+    user_sketch.start()
+
+    @window.event
+    def on_draw():
+        pass
+
+    def update(delta_time):
+        user_sketch.update()
+
+    pyglet.clock.schedule_interval(update, 1/settings.fps)
+    pyglet.clock.schedule_once(update, 0)
+
+    if arguments.timeout > 0:
+        pyglet.clock.schedule_once(window.exit_callback, arguments.timeout)
+        pyglet.app.run()
+    else:
+        pyglet.app.run()
 
 
 if __name__ == "__main__":
