+++ pyperlib/drawing/__init__.py
@@ -1,3 +1,10 @@
-from pyperlib.drawing.color import *
-from pyperlib.drawing.window import *
-from pyperlib.drawing.primitives import *
+# Color
+from pyperlib.drawing.color import Color, colors
+
+# Window
+from pyperlib.drawing.window import PyperWindow, context
+
+# Primitives
+from pyperlib.drawing.primitives import rectangle
+from pyperlib.drawing.primitives import clear, clear_screen
+from pyperlib.drawing.primitives import fill, fill_screen
