+++ pyperlib/settings.py
@@ -1,10 +1,12 @@
 
 
 class Settings:
-    def __init__(self, fps):
+    def __init__(self, fps, opengl_version):
         self.fps = fps
+        self.opengl_version = opengl_version
 
 
 settings = Settings(
+    opengl_version=300,
     fps=60
 )
